参照周りを修正 視点を分割するよう変更

+++ Culling3D/Culling3D.Layer.cpp
@@ -38,7 +38,7 @@ namespace Culling3D
 
 		int32_t ind = xind + yind * this->gridXCount + zind * this->gridXCount * this->gridYCount;
 
-		if (ind < 0 || ind >= grids.size()) return false;
+		if (ind < 0 || ind >= (int32_t)grids.size()) return false;
 
 		grids[ind].AddObject(o);
 
@@ -61,7 +61,7 @@ namespace Culling3D
 
 		int32_t ind = xind + yind * this->gridXCount + zind * this->gridXCount * this->gridYCount;
 
-		if (ind < 0 || ind >= grids.size()) return false;
+		if (ind < 0 || ind >= (int32_t) grids.size()) return false;
 
 		grids[ind].RemoveObject(o);
